Code quality changes + refactors
